some minor reworking of file naming
